sesion id y bg para analisis

+++ Alcachofita_2/Assets/Scripts/Telemetria/Eventos/DrawEndEvent.cs
@@ -7,7 +7,7 @@ public class DrawEndEvent : TrackerEvent
     public float clic_pos_x; // Coordenada X del clic respecto a la resolución de la pantalla de juego
     public float clic_pos_y; // Coordenada Y del clic respecto a la resolución de la pantalla de juego
 
-    public DrawEndEvent(int eventId, DateTime timestamp, int sessionId, float mouse_pos_x, float mouse_pos_y) : base("draw_start", "Gameplay", eventId, timestamp, sessionId) {
+    public DrawEndEvent(int eventId, DateTime timestamp, long sessionId, float mouse_pos_x, float mouse_pos_y) : base("draw_start", "Gameplay", eventId, timestamp, sessionId) {
         this.clic_pos_x = mouse_pos_x;
         this.clic_pos_y = mouse_pos_y;
     }
