+++ Alcachofita_2/Assets/Scripts/Telemetria/Eventos/LevelEndEvent.cs
@@ -6,7 +6,7 @@ public class LevelEndEvent : TrackerEvent {
     // True == nivel superado
     // False == nivel perdido
     public bool result;
-    public LevelEndEvent(int eventId, DateTime timestamp, int sessionId, byte levelid,  bool result) : base("level_end", "Gameplay", eventId, timestamp, sessionId) {
+    public LevelEndEvent(int eventId, DateTime timestamp, long sessionId, byte levelid,  bool result) : base("level_end", "Gameplay", eventId, timestamp, sessionId) {
         this.level_id = levelid;
         this.result = result;
     }
