+++ Alcachofita_2/Assets/Scripts/Telemetria/Eventos/LevelStartEvent.cs
@@ -5,7 +5,7 @@ using System;
 public class LevelStartEvent : TrackerEvent {
     // Número que corresponde al dibujo en el manual.
     public byte level_id;
-    public LevelStartEvent(int eventId, DateTime timestamp, int sessionId, byte levelid) : base("level_start", "Gameplay", eventId, timestamp, sessionId) {
+    public LevelStartEvent(int eventId, DateTime timestamp, long sessionId, byte levelid) : base("level_start", "Gameplay", eventId, timestamp, sessionId) {
         this.level_id = levelid;
     }
 }
