+++ Alcachofita_2/Assets/Scripts/Telemetria/Eventos/UIInteracionEvent.cs
@@ -9,7 +9,7 @@ public class UIInteractionEvent : TrackerEvent {
     public float clic_pos_x; // Coordenada X del clic respecto a la resolución de la pantalla de juego
     public float clic_pos_y; // Coordenada Y del clic respecto a la resolución de la pantalla de juego
 
-    public UIInteractionEvent(int eventId, DateTime timestamp, int sessionId, InteractionTarget target, float clicPosX, float clicPosY) : base("ui_interaction", "UI", eventId, timestamp, sessionId) { 
+    public UIInteractionEvent(int eventId, DateTime timestamp, long sessionId, InteractionTarget target, float clicPosX, float clicPosY) : base("ui_interaction", "UI", eventId, timestamp, sessionId) { 
         this.target_id = target;
         this.clic_pos_x = clicPosX;
         this.clic_pos_y = clicPosY;
