+++ Alcachofita_2/Assets/Scripts/Telemetria/Eventos/WindowBackgroundedEvent.cs
@@ -3,5 +3,5 @@ using System;
 // Puesta en segundo plano o minimización de la ventana.
 public class WindowBackgroundedEvent : TrackerEvent {
     // No tiene ningún atributo extra.
-    public WindowBackgroundedEvent(int eventId, DateTime timestamp, int sessionId) : base("window_backgrounded", "UI", eventId, timestamp, sessionId) { }
+    public WindowBackgroundedEvent(int eventId, DateTime timestamp, long sessionId) : base("window_backgrounded", "UI", eventId, timestamp, sessionId) { }
 }
