+++ Alcachofita_2/Assets/Scripts/Telemetria/Eventos/WindowForegroundedEvent.cs
@@ -3,5 +3,5 @@ using System;
 // Puesta en primer plano o maximización de la ventana.
 public class WindowForegroundedEvent : TrackerEvent {
     // No tiene ningún atributo extra.
-    public WindowForegroundedEvent(int eventId, DateTime timestamp, int sessionId) : base("window_foregrounded", "UI", eventId, timestamp, sessionId) { }
+    public WindowForegroundedEvent(int eventId, DateTime timestamp, long sessionId) : base("window_foregrounded", "UI", eventId, timestamp, sessionId) { }
 }
